ezyschooling/requirements/req.txt
amqp==2.6.1
astroid==2.9.3
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.8.2
billiard==3.6.4.0
black==22.1.0
bleach==4.1.0
boto3==1.21.9
botocore==1.24.9
cachetools==5.0.0
celery==4.4.0
certifi==2021.10.8
cffi==1.15.0
chardet==5.0.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==2.2.10
django-admin-rangefilter==0.6.3
django-allauth==0.41.0
django-autocomplete-light==3.5.1
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==2.2.0
django-celery-beat==1.6.0
django-ckeditor==5.9.0
django-cors-headers==3.2.1
django-crispy-forms==1.8.1
django-debug-toolbar==2.2
django-elasticsearch-dsl==7.1.1
django-elasticsearch-dsl-drf==0.20.9
django-extensions==2.2.6
django-filter==2.2.0
django-fixture-magic==0.1.5
django-guardian==2.2.0
django-import-export==2.0.1
django-js-asset==2.0.0
django-model-utils==4.2.0
django-nested-admin==3.2.4
django-nine==0.2.5
django-notifications-hq==1.5.0
django-post-office==3.6.0
django-redis==4.11.0
django-rest-auth==0.9.5
django-ses==2.6.1
django-storages==1.12.3
django-tables2==2.3.1
django-taggit==1.2.0
django-timezone-field==3.1
django-user-agents==0.4.0
django-webpush==0.3.2
djangorestframework==3.11.0
drf-yasg==1.17.0
elasticsearch==7.5.1
elasticsearch-dsl==7.1.0
entrypoints==0.3
environ==1.0
et-xmlfile==1.1.0
Faker==4.0.0
flake8==3.7.9
flower==0.9.3
furl==2.1.3
geographiclib==1.52
geopy==1.21.0
google-auth==2.6.0
google-auth-oauthlib==0.5.0
googlemaps==4.4.2
greenlet==1.1.2
gspread==3.6.0
html5lib==1.1
http-ece==1.1.0
idna==3.3
importlib-metadata==1.7.0
inflection==0.5.1
ipdb==0.13.9
ipython==7.32.0
isort==4.3.21
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.3
jmespath==0.10.0
jsonfield==3.1.0
kombu==4.6.11
lazy-object-proxy==1.7.1
logging-tree==1.9
logx==0.1.1579232358
lxml==4.8.0
Markdown==3.1.1
MarkupPy==1.14
MarkupSafe==2.1.0
matplotlib-inline==0.1.3
mccabe==0.6.1
migra==3.0.1652017683
mypy-extensions==0.4.3
numpy==1.21.5
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.9
orderedmultidict==1.0.1
packaging==21.3
pandas==1.1.5
parso==0.8.3
pathspec==0.9.0
pdf2image==1.16.0
pdf2img==0.1.2
pdfminer==20191125
pendulum==2.1.2
pickleshare==0.7.5
pikepdf==5.1.3
Pillow==9.0.1
platformdirs==2.5.1
prompt-toolkit==3.0.28
psycopg2-binary==2.8.4
py-vapid==1.8.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.21
pycryptodome==3.14.1
pyfcm==1.5.4
pyflakes==2.1.1
Pygments==2.5.2
pylint==2.12.2
pylint-django==2.0.13
pylint-plugin-utils==0.7
PyMuPDF==1.19.6
pyotp==2.4.1
pyparsing==3.0.7
PyPDF2==1.26.0
pytesseract==0.3.9
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.3
python-monkey-business==1.0.0
python3-openid==3.2.0
pytz==2021.3
pytzdata==2020.1
pywebpush==1.9.4
PyYAML==6.0
razorpay==1.2.0
redis==3.4.1
reportlab==3.6.6
requests==2.27.1
requests-oauthlib==1.3.1
results==0.1.1588569394
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.5.2
schemainspect==3.1.1652015119
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.39
sqlbag==0.1.1617247075
sqlparse==0.4.2
tablib==0.14.0
tabulate==0.8.10
taggit-selectize==2.7.1
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tornado==5.1.1
traitlets==5.1.1
typed-ast==1.5.2
typing_extensions==4.1.1
ua-parser==0.10.0
uritemplate==4.1.1
urllib3==1.26.8
user-agents==2.2.0
vine==1.3.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==0.16.1
whitenoise==5.0.1
wrapt==1.13.3
xhtml2pdf==0.2.4
xlrd==2.0.1
xlwt==1.3.0
zipp==3.7.0
